+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-16 11:07:38
-// Commit: feat(test): integrate custom telemetry logger in test
+// Updated at: 2026-05-18 12:36:08
+// Commit: chore(test): update dependencies and cleanup unused imports in test
 
 class testComponentState {
-  final String commitRef = '1201';
-  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
-  final DateTime timestamp = DateTime.parse('2026-05-16T11:07:38+05:00');
+  final String commitRef = '1240';
+  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
+  final DateTime timestamp = DateTime.parse('2026-05-18T12:36:08+05:00');
 }